+++ integrations/github/github/core/exporters/repository_exporter.py
@@ -4,21 +4,19 @@ from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE, RAW_ITEM
 from port_ocean.utils.cache import cache_iterator_result
 from loguru import logger
 from github.core.options import ListRepositoryOptions, SingleRepositoryOptions
-from github.clients.rest_client import GithubRestClient
+from github.clients.http.rest_client import GithubRestClient
 
 
 class RestRepositoryExporter(AbstractGithubExporter[GithubRestClient]):
-
     async def get_resource[
         ExporterOptionsT: SingleRepositoryOptions
     ](self, options: ExporterOptionsT) -> RAW_ITEM:
-
         endpoint = (
             f"{self.client.base_url}/repos/{self.client.organization}/{options['name']}"
         )
         response = await self.client.send_api_request(endpoint)
         logger.info(f"Fetched repository with identifier: {options['name']}")
-        return response.json()
+        return response
 
     @cache_iterator_result()
     async def get_paginated_resources[