+++ integrations/github/main.py
@@ -2,23 +2,18 @@ from typing import cast
 from loguru import logger
 from port_ocean.context.ocean import ocean
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
-from github.clients.client_factory import create_github_client
+from github.clients.client_factory import (
+    GitHubAuthenticatorFactory,
+    create_github_client,
+)
 from github.clients.utils import integration_config
 from github.helpers.utils import ObjectKind
 from github.webhook.events import WEBHOOK_CREATE_EVENTS
 from github.webhook.webhook_processors.repository_webhook_processor import (
     RepositoryWebhookProcessor,
 )
-from github.webhook.webhook_processors.deployment_webhook_processor import (
-    DeploymentWebhookProcessor,
-)
 from github.webhook.webhook_client import GithubWebhookClient
-from github.core.options import ListDeploymentsOptions, ListEnvironmentsOptions
-from port_ocean.utils.async_iterators import stream_async_iterators_tasks
-
 from github.core.exporters.repository_exporter import RestRepositoryExporter
-from github.core.exporters.deployment_exporter import RestDeploymentExporter
-from github.core.exporters.environment_exporter import RestEnvironmentExporter
 from github.core.options import ListRepositoryOptions
 from typing import TYPE_CHECKING
 from port_ocean.context.event import event
@@ -40,8 +35,16 @@ async def on_start() -> None:
     if not base_url:
         return
 
+    authenticator = GitHubAuthenticatorFactory.create(
+        organization=ocean.integration_config["github_organization"],
+        github_host=ocean.integration_config["github_host"],
+        token=ocean.integration_config.get("github_token"),
+        app_id=ocean.integration_config.get("github_app_id"),
+        private_key=ocean.integration_config.get("github_app_private_key"),
+    )
+
     client = GithubWebhookClient(
-        **integration_config(),
+        **integration_config(authenticator),
         webhook_secret=ocean.integration_config["webhook_secret"],
     )
 
@@ -58,7 +61,7 @@ async def resync_repositories(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     exporter = RestRepositoryExporter(rest_client)
 
     port_app_config = cast("GithubPortAppConfig", event.port_app_config)
-    options = ListRepositoryOptions(type=port_app_config.repository_visibility_filter)
+    options = ListRepositoryOptions(type=port_app_config.repository_type)
 
     async for repositories in exporter.get_paginated_resources(options):
         yield repositories