+++ integrations/github/tests/exporters/test_repository_exporter.py
@@ -1,13 +1,6 @@
 from typing import Any, AsyncGenerator
-from port_ocean.core.handlers.port_app_config.models import (
-    ResourceConfig,
-    Selector,
-    PortResourceConfig,
-    EntityMapping,
-    MappingsConfig,
-)
 import pytest
-from unittest.mock import patch, MagicMock
+from unittest.mock import AsyncMock, patch, MagicMock
 import httpx
 from github.core.exporters.repository_exporter import (
     RestRepositoryExporter,
@@ -15,7 +8,7 @@ from github.core.exporters.repository_exporter import (
 from integration import GithubPortAppConfig
 from port_ocean.context.event import event_context
 from github.core.options import SingleRepositoryOptions, ListRepositoryOptions
-from github.clients.rest_client import GithubRestClient
+from github.clients.http.rest_client import GithubRestClient
 
 
 TEST_REPOS = [
@@ -34,34 +27,8 @@ TEST_REPOS = [
 ]
 
 
-@pytest.fixture
-def mock_port_app_config() -> GithubPortAppConfig:
-    return GithubPortAppConfig(
-        delete_dependent_entities=True,
-        create_missing_related_entities=False,
-        repository_visibility_filter="all",
-        resources=[
-            ResourceConfig(
-                kind="repository",
-                selector=Selector(query="true"),
-                port=PortResourceConfig(
-                    entity=MappingsConfig(
-                        mappings=EntityMapping(
-                            identifier=".full_name",
-                            title=".name",
-                            blueprint='"githubRepository"',
-                            properties={},
-                        )
-                    )
-                ),
-            )
-        ],
-    )
-
-
 @pytest.mark.asyncio
 class TestRestRepositoryExporter:
-
     async def test_get_resource(self, rest_client: GithubRestClient) -> None:
         # Create a mock response
         mock_response = MagicMock(spec=httpx.Response)
@@ -71,8 +38,9 @@ class TestRestRepositoryExporter:
         exporter = RestRepositoryExporter(rest_client)
 
         with patch.object(
-            rest_client, "send_api_request", return_value=mock_response
+            rest_client, "send_api_request", new_callable=AsyncMock
         ) as mock_request:
+            mock_request.return_value = mock_response.json()
             repo = await exporter.get_resource(SingleRepositoryOptions(name="repo1"))
 
             assert repo == TEST_REPOS[0]
@@ -84,7 +52,6 @@ class TestRestRepositoryExporter:
     async def test_get_paginated_resources(
         self, rest_client: GithubRestClient, mock_port_app_config: GithubPortAppConfig
     ) -> None:
-
         # Create an async mock to return the test repos
         async def mock_paginated_request(
             *args: Any, **kwargs: Any
@@ -94,10 +61,9 @@ class TestRestRepositoryExporter:
         with patch.object(
             rest_client, "send_paginated_request", side_effect=mock_paginated_request
         ) as mock_request:
-
             async with event_context("test_event"):
                 options = ListRepositoryOptions(
-                    type=mock_port_app_config.repository_visibility_filter
+                    type=mock_port_app_config.repository_type
                 )
                 exporter = RestRepositoryExporter(rest_client)